Fix double instantiation of plugin screens
Reat were not sure on if should update the serverboards component after
loading a plugin screen (answer is not), so it was doing it always.

This caused to load th eplugin twice, once without the proper existing
element, and xterm.js just adds one if needed.

Forcing not update unless really necessary (!= section, subsection or
location state) fixes this.