Added spice support via spice-html5 client
Works as expected!

It uses a spice.html/js as a shim to be able to load it in a iframe.

The iframe trick is not really required, but allows for easy garbage
collect of all memory and resources from the iframe as soon as we leave
the screen.

It is prefered as spice support can be a lot of code that may interfere
with normal serverboards functionalities.