Now uses the serverboard as prefix for the data key
This means that the secrets are compartmeterized into the serverboards.

BUT THIS IS NOT SECURE COMPARTMENTIZATION.

There are no security in place to be able to get the data from the wrong
serverboard. Access to this plugin is to all serverboards secrets. But
you need the key to really see it.