WIP: New managed node_exporter, and managed prometheus
Both install the node_exporter on the remote SSH node, and sets it up
(only systemd now), or locally and run it along with serverboards.

Whenever an agent or a enode_exporter service is updated, it updates the
managed prometheus services, which allows to have up to date data for
all nodes at prometheus.

With this commit al lbasic functionallity works, but there are more
cases to be covered.