Soved misterious reordering fail bug on ssh.popen
When on write_to_ssh reading, more "ready" may arrive, wich will make
the new calls async but answered before due to the way the serverboards
python rpc bindings work. This makes each call get the proper data they
asked, but in diferent orders, so that the second call may get the data
before the first call does. This makes it write the data in the worng
order.

With this hack, this prevents the second call until the first is
answered. A possibe solution would be to disallow reentry on events
until previous event calls to the same function has returned, for
example queing for later to do a delayed event signaling.