Fixed tests to pass having mustache as templating engine
This made a side effect that some expected behaviours were not met, but
not important on client side. On server side will have to create custom
mustache interpreter if we want more features.

Now it removes unknown tags, and does not have escaping (which actually
was using unknown tags).

Before it was leaving them as is, in the final text.